Achieved the detection of tags
Change filter to get external squares  because in image with long
distance, it just detect external , not internals.
Added a method to remove this border from squares already detected.
It's a estimation based of the length of square, previously known.
Fixed selection of good increment for height and width, in order to
detect the circles inside the tag.
Added operation erode to bImage and gImage in cutSquares.