Change app icon and added support to new tags
Added new icon for the app.
Change name of folder where the image are stored to "TagRecognizerApp".
Added new kind of tags: 2x4.
Fix problem related with even number of rows.
Adjust value of borders, in order to remove them.
Add global variable with the path of the folder in JNI